+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Fully-loaded blueprint book with all blueprints in creative menu
+### Changed
+- Doubled the inventory size of the blueprint book
+
 ## [2.0.2] - 2020-07-29
 1.16.1-only bug fix
 ### Added